common code moved to VisualElementRenderer
