+++ Xamarin.Forms.Platform.UAP/DatePickerRenderer.cs
@@ -3,18 +3,16 @@ using System.ComponentModel;
 using System.Linq;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Xamarin.Forms.Internals;
 
 namespace Xamarin.Forms.Platform.UWP
 {
-	public class DatePickerRenderer : ViewRenderer<DatePicker, Windows.UI.Xaml.Controls.DatePicker>
+	public class DatePickerRenderer : ViewRenderer<DatePicker, Windows.UI.Xaml.Controls.DatePicker>, ITabStopOnDescenants
 	{
 		Brush _defaultBrush;
 		bool _fontApplied;
 		FontFamily _defaultFontFamily;
-		FocusNavigationDirection focusDirection;
 
 		protected override void Dispose(bool disposing)
 		{
@@ -22,8 +20,6 @@ namespace Xamarin.Forms.Platform.UWP
 			{
 				Control.DateChanged -= OnControlDateChanged;
 				Control.Loaded -= ControlOnLoaded;
-				Control.GotFocus -= OnGotFocus;
-				Control.GettingFocus -= OnGettingFocus;
 			}
 
 			base.Dispose(disposing);
@@ -39,8 +35,6 @@ namespace Xamarin.Forms.Platform.UWP
 					SetNativeControl(picker);
 					Control.Loaded += ControlOnLoaded;
 					Control.DateChanged += OnControlDateChanged;
-					Control.GotFocus += OnGotFocus;
-					Control.GettingFocus += OnGettingFocus;
 				}
 				else
 				{
@@ -54,20 +48,6 @@ namespace Xamarin.Forms.Platform.UWP
 			}
 
 			base.OnElementChanged(e);
-		}
-
-		protected override void UpdateTabStop()
-		{
-			base.UpdateTabStop();
-			Control?.GetChildren<Control>().ForEach(c => c.IsTabStop = Element.IsTabStop);
-		}
-
-		void OnGettingFocus(UIElement sender, GettingFocusEventArgs args) => focusDirection = args.Direction;
-
-		void OnGotFocus(object sender, RoutedEventArgs e)
-		{
-			if (e.OriginalSource == Control)
-				FocusManager.TryMoveFocus(focusDirection);
 		}
 
 		void ControlOnLoaded(object sender, RoutedEventArgs routedEventArgs)
