+++ Xamarin.Forms.Platform.UAP/ITabStopOnDescenants.cs
@@ -1,0 +1,6 @@
+﻿namespace Xamarin.Forms.Platform.UWP
+{
+	public interface ITabStopOnDescenants
+	{
+	}
+}
