+++ Xamarin.Forms.Platform.UAP/TimePickerRenderer.cs
@@ -1,21 +1,18 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Xamarin.Forms.Internals;
 
 namespace Xamarin.Forms.Platform.UWP
 {
-	public class TimePickerRenderer : ViewRenderer<TimePicker, Windows.UI.Xaml.Controls.TimePicker>
+	public class TimePickerRenderer : ViewRenderer<TimePicker, Windows.UI.Xaml.Controls.TimePicker>, ITabStopOnDescenants
 	{
 		Brush _defaultBrush;
 		bool _fontApplied;
 		FontFamily _defaultFontFamily;
-		FocusNavigationDirection focusDirection;
 
 		protected override void Dispose(bool disposing)
 		{
@@ -23,8 +20,6 @@ namespace Xamarin.Forms.Platform.UWP
 			{
 				Control.TimeChanged -= OnControlTimeChanged;
 				Control.Loaded -= ControlOnLoaded;
-				Control.GotFocus -= OnGotFocus;
-				Control.GettingFocus -= OnGettingFocus;
 			}
 
 			base.Dispose(disposing);
@@ -44,8 +39,6 @@ namespace Xamarin.Forms.Platform.UWP
 
 					Control.TimeChanged += OnControlTimeChanged;
 					Control.Loaded += ControlOnLoaded;
-					Control.GotFocus += OnGotFocus;
-					Control.GettingFocus += OnGettingFocus;
 				}
 				else
 				{
@@ -55,20 +48,6 @@ namespace Xamarin.Forms.Platform.UWP
 				UpdateTime();
 				UpdateFlowDirection();
 			}
-		}
-
-		protected override void UpdateTabStop()
-		{
-			base.UpdateTabStop();
-			Control?.GetChildren<Control>().ForEach(c => c.IsTabStop = Element.IsTabStop);
-		}
-
-		void OnGettingFocus(UIElement sender, GettingFocusEventArgs args) => focusDirection = args.Direction;
-
-		void OnGotFocus(object sender, RoutedEventArgs e)
-		{
-			if (e.OriginalSource == Control)
-				FocusManager.TryMoveFocus(focusDirection);
 		}
 
 		void ControlOnLoaded(object sender, RoutedEventArgs routedEventArgs)
